lotto_114126_package_updated/lotto_114126_v2_upd/requirements.txt
torch>=2.0
stable-baselines3>=2.3.0
pandas
numpy
matplotlib
gymnasium
yaml
